Consolidate all modal components
