--- ui/src/Modal/AvailableModals.tsx
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2022 Ford Motor Company
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-export enum AvailableModals {
-    CREATE_PRODUCT,
-    CREATE_PRODUCT_OF_PRODUCT_TAG,
-    CREATE_PRODUCT_OF_LOCATION,
-    EDIT_PRODUCT,
-    CREATE_PERSON,
-    EDIT_PERSON,
-    CREATE_ASSIGNMENT,
-    ASSIGNMENT_EXISTS_WARNING,
-    MY_ROLES_MODAL,
-    CREATE_SPACE,
-    EDIT_SPACE,
-    DELETE_SPACE,
-    SHARE_SPACE_ACCESS,
-    GRANT_EDIT_ACCESS_CONFIRMATION,
-    MY_LOCATION_TAGS,
-    MY_PRODUCT_TAGS,
-    MY_PERSON_TAGS,
-    TRANSFER_OWNERSHIP,
-    DELETE_SPACE_NO_EDITORS
-}
