+++ ui/src/Modal/ConfirmationModal/ConfirmationModal.scss
@@ -1,0 +1,129 @@
+/*
+ * Copyright (c) 2022 Ford Motor Company
+ * All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@import 'Styles/Colors.scss';
+@import 'Styles/Breakpoints.scss';
+@import '../Form.scss';
+
+.modalBackground {
+  position: fixed;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(black, 0.8);
+  z-index: 9;
+  overflow-y: auto;
+}
+
+.modalContents {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin: 32px auto;
+  min-height: calc(100% - 64px);
+  max-width: 500px;
+  width: auto;
+}
+
+.modalCard {
+  display: flex;
+  flex-direction: column;
+  text-align: center;
+  background: $white;
+  width: 494px;
+  border-radius: 6px;
+  padding: 32px 0;
+  margin: 0.5rem 0;
+
+  @include phone {
+    width: 100vw;
+    box-sizing: border-box;
+  }
+
+  &[hidden] {
+    overflow: hidden;
+    max-height: 65px;
+
+    @include phone {
+      max-height: initial;
+    }
+  }
+}
+
+.modalCardBanner {
+  display: flex;
+  align-items: center;
+  padding: 0 0 32px 0;
+  margin-left: 36px;
+  margin-right: 36px;
+
+  .modalTitle {
+    flex: 2;
+    font-size: 20px;
+    color: $purple-title;
+    font-weight: bold;
+    margin: 0 0 0 16px;
+  }
+
+  button.closeButton {
+    cursor: pointer;
+    border-radius: 50%;
+    background: none;
+    color: black;
+    padding: 0;
+    letter-spacing: normal;
+  }
+}
+
+.confirmationModalContent {
+  font-size: 14px;
+  color: $gray-1;
+  padding: 0 52px;
+  text-align: left;
+  line-height: 16px;
+}
+
+.confirmationModalControls.secondaryButtonContainer {
+  justify-content: space-between;
+  width: unset;
+  align-self: unset;
+}
+
+.confirmationControlButtons {
+  margin-top: 32px;
+  justify-content: space-evenly;
+  width: 50%;
+  align-self: center;
+}
+
+.cancelAndArchiveContainer {
+  justify-content: flex-start;
+}
+
+.cancelAndArchiveContainer.secondaryButtonContainer {
+  display: flex;
+  flex-direction: row;
+  min-width: 218px;
+  justify-content: space-between;
+}
+
+.hiddenInputField {
+  display: none;
+}
