+++ ui/src/Modal/Modal.scss
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Ford Motor Company
+ * Copyright (c) 2022 Ford Motor Company
  * All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,115 +15,72 @@
  * limitations under the License.
  */
 
-@import '../Styles/Colors.scss';
-@import '../Styles/Breakpoints.scss';
+@import 'Styles/main';
 @import './Form.scss';
 
-.modalBackground {
-    position: fixed;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(black, 0.8);
-    z-index: 9;
-    overflow-y: auto;
+.modal-container,
+.modal-overlay {
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
 }
 
-.modalContents {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    margin: 32px auto;
-    min-height: calc(100% - 64px);
-    max-width: 500px;
-    width: auto;
-}
+.modal-container {
+  display: flex;
+  z-index: 5;
 
-.modalCard {
-    display: flex;
-    flex-direction: column;
-    text-align: center;
-    background: $white;
-    width: 494px;
-    border-radius: 6px;
-    padding: 32px 0;
-    margin: 0.5rem 0;
-
-    @include phone {
-        width: 100vw;
-        box-sizing: border-box;
-    }
-
-    &[hidden] {
-        overflow: hidden;
-        max-height: 65px;
-
-        @include phone {
-            max-height: initial;
-        }
-    }
-}
-
-.modalCardBanner {
-    display: flex;
-    align-items: center;
-    padding: 0 0 32px 0;
-    margin-left: 36px;
-    margin-right: 36px;
-
-    .modalTitle {
-        flex: 2;
-        font-size: 20px;
-        color: $purple-title;
-        font-weight: bold;
-        margin: 0 0 0 16px;
-    }
-
-    button.closeButton {
-        cursor: pointer;
-        border-radius: 50%;
-        background: none;
-        color: black;
-        padding: 0;
-        letter-spacing: normal;
-    }
-}
-
-.confirmationModalContent {
-    font-size: 14px;
-    color: $gray-1;
-    padding: 0 52px;
-    text-align: left;
-    line-height: 16px;
-}
-
-.confirmationModalControls.secondaryButtonContainer {
-    justify-content: space-between;
-    width: unset;
-    align-self: unset;
-}
-
-.confirmationControlButtons {
-    margin-top: 32px;
-    justify-content: space-evenly;
-    width: 50%;
-    align-self: center;
-}
-
-.cancelAndArchiveContainer {
-    justify-content: flex-start;
-}
-
-.cancelAndArchiveContainer.secondaryButtonContainer {
-    display: flex;
-    flex-direction: row;
-    min-width: 218px;
-    justify-content: space-between;
-}
-
-.hiddenInputField {
+  &[aria-hidden='true'] {
     display: none;
+  }
+
+  .modal-overlay {
+    background-color: rgba(black, 0.8);
+  }
+
+  .modal-content {
+    background-color: white;
+    border-radius: 6px;
+    width: 100%;
+    margin: auto;
+    padding: 2rem;
+
+    color: $gray-1;
+    font-family: 'Quicksand', sans-serif;
+    font-weight: normal;
+    font-size: 1rem;
+
+    z-index: 2;
+    position: relative;
+
+    @include tablet {
+      width: unset;
+    }
+
+    .modal-title {
+      text-align: center;
+      font-size: 20px;
+      color: $purple-title;
+      font-weight: bold;
+      margin: 0 2rem 2rem;
+    }
+
+    .modal-close-button {
+      cursor: pointer;
+      border-radius: 50%;
+      background: none;
+      color: black;
+      padding: 0;
+      letter-spacing: normal;
+      margin: 0.5rem;
+      position: absolute;
+      right: 0;
+      top: 0;
+
+      @include tablet {
+        margin: 2rem;
+      }
+    }
+  }
 }
