+++ ui/src/Modal/MultiModalCardBanner/MultiModalCardBanner.tsx
@@ -17,7 +17,7 @@
 
 import {JSX} from '@babel/types';
 import React from 'react';
-import {CloseModalButton, ModalTitle} from './ModalCardBanner';
+import {CloseModalButton, ModalTitle} from 'Modal/ModalCardBanner/ModalCardBanner';
 
 import './MultiModalCardBanner.scss';
