--- ui/src/NewModal/Modal.scss
@@ -1,85 +1,0 @@
-/*
- * Copyright (c) 2022 Ford Motor Company
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@import 'Styles/main';
-
-.modal-container,
-.modal-overlay {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-}
-
-.modal-container {
-  display: flex;
-  z-index: 5;
-
-  &[aria-hidden='true'] {
-    display: none;
-  }
-
-  .modal-overlay {
-    background-color: rgba(black, 0.8);
-  }
-
-  .modal-content {
-    background-color: white;
-    border-radius: 6px;
-    width: 100%;
-    margin: auto;
-    padding: 2rem;
-
-    color: $gray-1;
-    font-family: 'Quicksand', sans-serif;
-    font-weight: normal;
-    font-size: 1rem;
-
-    z-index: 2;
-    position: relative;
-
-    @include tablet {
-      width: unset;
-    }
-
-    .modal-title {
-      text-align: center;
-      font-size: 20px;
-      color: $purple-title;
-      font-weight: bold;
-      margin: 0 2rem 2rem;
-    }
-
-    .modal-close-button {
-      cursor: pointer;
-      border-radius: 50%;
-      background: none;
-      color: black;
-      padding: 0;
-      letter-spacing: normal;
-      margin: 0.5rem;
-      position: absolute;
-      right: 0;
-      top: 0;
-
-      @include tablet {
-        margin: 2rem;
-      }
-    }
-  }
-}
