+++ ui/src/Routes.tsx
@@ -30,7 +30,7 @@ import AnimatedImageSrc from './Assets/404.gif';
 import errorImageSrc from './Assets/403.png';
 import Header from './Header/Header';
 import AnnouncementBanner from './Header/AnnouncementBanner';
-import Modal from 'NewModal/Modal';
+import Modal from 'Modal/Modal';
 
 export const dashboardUrl = '/user/dashboard';
 const notFoundUrl = '/error/404';
